backend/requirements.txt
flask
flask-cors
firebase-admin
python-dotenv
groq